Process performed only if the corresponding file exists.

+++ build/batik-patch.sh
@@ -10,17 +10,26 @@ echo "Fixing BATIK dependencies for CS-Studio ${CSSVER}..."
 echo "--------------------------------------------------"
 
 echo "Patching Window product:"
-zip -d cs-studio-ess-${CSSVER}-win32.win32.x86_64.zip "plugins/org.apache.batik*1.8*.jar"; \
+if [ -f "cs-studio-ess-${CSSVER}-win32.win32.x86_64.zip" ]
+then
+  zip -d cs-studio-ess-${CSSVER}-win32.win32.x86_64.zip "plugins/org.apache.batik*1.8*.jar"; \
+fi
 
 echo "Patching MacOS X product:"
-zip -d cs-studio-ess-${CSSVER}-macosx.cocoa.x86_64.zip "ESS CS-Studio.app/Contents/Eclipse/plugins/org.apache.batik*1.8*.jar"; \
+if [ -f "cs-studio-ess-${CSSVER}-macosx.cocoa.x86_64.zip" ]
+then
+  zip -d cs-studio-ess-${CSSVER}-macosx.cocoa.x86_64.zip "ESS CS-Studio.app/Contents/Eclipse/plugins/org.apache.batik*1.8*.jar"; \
+fi
 
 echo "Patching Linux product:"
-mkdir cs-studio-ess-${CSSVER}-linux.gtk.x86_64
-cd cs-studio-ess-${CSSVER}-linux.gtk.x86_64
-tar -zxf ../cs-studio-ess-${CSSVER}-linux.gtk.x86_64.tar.gz
-find ./cs-studio/plugins -type f -iname 'org.apache.batik*1.8*.jar' -delete -print
-rm -f ../cs-studio-ess-${CSSVER}-linux.gtk.x86_64.tar.gz
-tar -zcf ../cs-studio-ess-${CSSVER}-linux.gtk.x86_64.tar.gz cs-studio
-cd ..
-rm -fr cs-studio-ess-${CSSVER}-linux.gtk.x86_64
+if [ -f "cs-studio-ess-${CSSVER}-linux.gtk.x86_64.tar.gz" ]
+then
+  mkdir cs-studio-ess-${CSSVER}-linux.gtk.x86_64
+  cd cs-studio-ess-${CSSVER}-linux.gtk.x86_64
+  tar -zxf ../cs-studio-ess-${CSSVER}-linux.gtk.x86_64.tar.gz
+  find ./cs-studio/plugins -type f -iname 'org.apache.batik*1.8*.jar' -delete -print
+  rm -f ../cs-studio-ess-${CSSVER}-linux.gtk.x86_64.tar.gz
+  tar -zcf ../cs-studio-ess-${CSSVER}-linux.gtk.x86_64.tar.gz cs-studio
+  cd ..
+  rm -fr cs-studio-ess-${CSSVER}-linux.gtk.x86_64
+fi
